gql_accounts_service/pyproject.toml
# TODO: UPDATE THIS PROJECT
[tool.poetry]
name = "knowledge-graphql-accounts-service"
version = "1.0.0"
description = ""
authors = [
    "Jorgen Ader <jorgen@thorgate.eu>"
]


[tool.poetry.dependencies]
python = "^3.8"
boto3 = "==1.9.*"
Django = "==2.*"
django-filter = "==2.*"
graphene-django = "==2.*"
graphene-federation = "==0.*"
django-graphql-jwt = "==0.*"
djangorestframework = "==3.*"
django-crispy-forms = "==1.7.*"
django-environ = "*"
django-redis = "==4.9.*"
django-cors-headers = "==3.*"
django-storages = "==1.7.*"
gunicorn = "==19.*"
uvicorn = "==0.*"
psycopg2-binary = "==2.8.*"
pyyaml = "==5.*"
sentry-sdk = "==0.13.*"
python-json-logger = ">= 0.1.*,<3.0"
redis = "==3.2.*"
tg-utils = { version = "*", extras =  ["model-hash", "health_check"] }
tg-react = "*"


[tool.poetry.dev-dependencies]
black = { version = "^19.10b0", python = "^3.8" }
django-debug-toolbar = "*"
coverage = "*"
flake8 = "==3.5.*"
isort = "==4.*"
pytest = "==4.*"
pytest-django = "*"
pytest-cov = "*"
prospector = "==1.1.6.2"
pycodestyle = "*"
pylint = "<2.3"
pylint-django = "*"
graphene-django-optimizer = "0.6.*"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | docs
  | fabfile.py
  | local.py
  | local_test.py
)
'''

[tool.isort]
skip = 'settings/local.py,settings/local_test.py,fabfile.py,project_light_keepers/wsgi.py,project_light_keepers/asgi.py'
skip_glob="node_modules"
line_length = '88'
atomic = 'true'
multi_line_output = '3'
include_trailing_comma = 'true'
known_django = "django"
known_pytest = "pytest"
known_first_party = "accounts,project_light_keepers"
lines_between_types = '0'
sections = "FUTURE,STDLIB,DJANGO,PYTEST,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
lines_after_imports = '2'
force_alphabetical_sort_within_sections = '1'
